refactor(shared): move components into their own folder
