+++ src/shared/loader/index.js
@@ -1,0 +1,1 @@
+export { Loader } from './loader.js'
