+++ src/shared/loader/loader.js
